requirements.txt
flask~=2.0.2
pycryptodome~=3.12.0
schema~=0.7.5
pymongo~=4.0